Allocator implementation is now separated from kalloc/free functions
